backend/requirements.txt
# Django核心框架
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
djangorestframework-simplejwt==5.3.0

# 数据库
psycopg2-binary==2.9.7  # PostgreSQL适配器
dj-database-url==2.1.0  # 数据库URL解析

# 缓存和任务队列
redis==5.0.1
django-redis==5.4.0
celery==5.3.4

# AI/ML库
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.25.2
joblib==1.3.2

# Google AI API
google-generativeai==0.3.2

# 数据处理和分析
matplotlib==3.8.2
seaborn==0.13.0

# 文件处理
Pillow==10.1.0
python-magic==0.4.27

# 环境变量管理
python-decouple==3.8
python-dotenv==1.0.0

# API文档
drf-spectacular==0.26.5

# 开发工具
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# 测试
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# 代码质量
flake8==6.1.0
black==23.11.0
isort==5.12.0

# 安全
django-ratelimit==4.1.0
django-guardian==2.4.0

# 日志
structlog==23.2.0

# 时间处理
pytz==2023.3

# HTTP客户端
requests==2.31.0

# 字符串处理
python-slugify==8.0.1

# 分页
django-rest-framework-pagination==0.1.1